requirements.txt
pandas
numpy
streamlit
yfinance
mplfinance
requests
python-dotenv
